support multiple singing key types
Using the new support in java-http-signature 4.0.0, support DSA and
ECDSA keys.  A user can provide a key and the library will figure out
it's type and use sensible defaults.

As part of this I had to tease apart how `ThreadLocalSigner` objects
are managed by centralizing their creation in `MantaClient`.  It was
too easy to end up with multiple ones with different configuration
which only accidental worked when everything was RSA.  While several
class changed, the public api of `MantaClient` did not.