detect last MPU part to maintain min size
Previously the buffer and HAMC bytes were always unconditionally
written as a final part.  This meant that if the last user generated
part was less than the minimum part size, two parts below the minimum
would be uploaded.

To avoid this, keep track of the bytes for uploaded for each part. If
less than the minimum bytes have been uploaded then this must be the
last part (otherwise there would be an error on completion).  If that
is the case append the remaining cipher bytes to the last part, and do
*not* create an additional part during completion.

ref #218