make cli: get-file suitable for streaming
Instead of reading the entire object into a string and running out of
memory, write out bytes as they are downloaded.  This also makes it
easy to stream binary content to a local file.

Add a nascent info command to retain a way for the cli to show
http headers.

ref #312