Provide Unshaded Artifact (#327)
Split the primary client module so that the core library can be published separately from the java-manta-client artifact. After this change all code which used to live inside java-manta-client was moved to java-manta-client-unshaded but the shading plugin configuration will continue to produce an artifact for the java-manta-client artifact. Documentation was added for users of the -unshaded variant.