Updating from the comments left.
Created another delete method that takes an integer, and if the integer
is null it just does the normal delete op. Else it uses the strategy.

Added a sentinel value to the strategy vs treating all of the negative
values the same.

Updated the test cases to reflect new methods, and added a test case for
numbers smaller than the sentinel value.