Fix/425 (#431)
* Adding the Prune Empty Parent Directory as per issues 425.

Provide a way to automatically prune empty parent directories

* Adding changes for Issue 425.

* Reverting some accidental format changes.

* Changing the limit test, to test for negative numbers and not just -1.

* Updating author & since comments.

* Fixing author comments, changed the wrong one so changing that one back,
and changing the correct one

* Updating from the comments left.

Created another delete method that takes an integer, and if the integer
is null it just does the normal delete op. Else it uses the strategy.

Added a sentinel value to the strategy vs treating all of the negative
values the same.

Updated the test cases to reflect new methods, and added a test case for
numbers smaller than the sentinel value.

* Updating the useage.md.

* Updating the USAGE.md.

* Addressing issues found in code review.

* Adding a changelog entry.

* Adding the entry to the changelog in the correct place. After looking at
this for a second I noticed that I put the entry in the wrong place.