MANTA-4201 track origin of mantaKeyPath in chained config
It was possible for a config to get into an illegal state, with both
mantaKeyPath and mantaKeyContent. This was due to the provenance of
the mantaKeyPath not being propagated through the
mantaKeyPathSetOnlyByDefaults field. This change uses a hack to check
that fields value when chained configs are added to chained configs.